Fix issues that break MSVC strict compilation
- Add an explicit public default constructor to ComponentStorage
  to avoid "no appropriate default constructor available" errors
- Explicitly define the static inline ComponentStorage member
  named storage, instead of only declaring it, to avoid
  "use of undefined type 'sp::ecs::ComponentStorage<...>'" errors
- Use _WIN32 instead of WIN32 in preprocessor ifdefs; MSVC 18
  doesn't appear to define WIN32.
- Explicitly instantiate multiplayerReplicationFunctions templates.

+++ src/container/sparseset.h
@@ -43,7 +43,7 @@ public:
             free_dense = new_free;
         } else {
             // Append to the data
-            sparse[index] = dense.size();
+            sparse[index] = static_cast<uint32_t>(dense.size());
             dense.push_back(index);
             data.emplace_back(value);
         }
@@ -67,7 +67,7 @@ public:
             free_dense = new_free;
         } else {
             // Append to the data
-            sparse[index] = dense.size();
+            sparse[index] = static_cast<uint32_t>(dense.size());
             dense.push_back(index);
             data.emplace_back(std::move(value));
         }
